+++ tests/TestCase/Extension/ListTableExtensionTest.php
@@ -54,7 +54,9 @@ class ListTableExtensionTest extends TestCase
         $expected = implode("\n", [
             '<table>',
             '  <caption>Quarterly results</caption>',
-            '  <thead><tr><th scope="col">Region</th><th scope="col">Notes</th></tr></thead>',
+            '  <thead>',
+            '    <tr><th scope="col">Region</th><th scope="col">Notes</th></tr>',
+            '  </thead>',
             '  <tbody>',
             '    <tr><td>EMEA</td><td>Strong quarter.</td></tr>',
             '  </tbody>',
@@ -164,7 +166,9 @@ class ListTableExtensionTest extends TestCase
 
         $expected = implode("\n", [
             '<table>',
-            '  <thead><tr><th scope="col">Region</th><th scope="col">Notes</th></tr></thead>',
+            '  <thead>',
+            '    <tr><th scope="col">Region</th><th scope="col">Notes</th></tr>',
+            '  </thead>',
             '  <tbody>',
             '    <tr><td>EMEA</td><td>ok</td></tr>',
             '  </tbody>',
@@ -214,7 +218,9 @@ class ListTableExtensionTest extends TestCase
         // The whole header row and the first column are all <th scope="col">.
         $expected = implode("\n", [
             '<table>',
-            '  <thead><tr><th scope="col">Metric</th><th scope="col">Q1</th><th scope="col">Q2</th></tr></thead>',
+            '  <thead>',
+            '    <tr><th scope="col">Metric</th><th scope="col">Q1</th><th scope="col">Q2</th></tr>',
+            '  </thead>',
             '  <tbody>',
             '    <tr><th scope="row">EMEA</th><td>1.0</td><td>1.2</td></tr>',
             '  </tbody>',
@@ -246,7 +252,10 @@ class ListTableExtensionTest extends TestCase
             '    <tr><th scope="row">Region</th><td>Q1</td></tr>',
             '    <tr><th scope="row">EMEA</th><td>10</td></tr>',
             '  </tbody>',
-            '  <tfoot><tr><th scope="row">Region</th><td>Q1</td></tr><tr><th scope="row">EMEA</th><td>10</td></tr></tfoot>',
+            '  <tfoot>',
+            '    <tr><th scope="row">Region</th><td>Q1</td></tr>',
+            '    <tr><th scope="row">EMEA</th><td>10</td></tr>',
+            '  </tfoot>',
             '</table>',
         ]);
         $this->assertSame($expected, $this->render($djot));
@@ -495,7 +504,9 @@ class ListTableExtensionTest extends TestCase
         $expected = implode("\n", [
             '<table>',
             '  <caption>Sales</caption>',
-            '  <thead><tr><th scope="col">Region</th><th scope="col">Q1</th><th scope="col">Q2</th></tr></thead>',
+            '  <thead>',
+            '    <tr><th scope="col">Region</th><th scope="col">Q1</th><th scope="col">Q2</th></tr>',
+            '  </thead>',
             '  <tbody>',
             '    <tr><td rowspan="2">EMEA</td><td>10</td><td>12</td></tr>',
             '    <tr><td>14</td><td>16</td></tr>',
@@ -566,7 +577,9 @@ class ListTableExtensionTest extends TestCase
 
         $expected = implode("\n", [
             '<table>',
-            '  <thead><tr><th scope="col">A</th><th scope="col">B</th><th scope="col">C</th></tr></thead>',
+            '  <thead>',
+            '    <tr><th scope="col">A</th><th scope="col">B</th><th scope="col">C</th></tr>',
+            '  </thead>',
             '  <tbody>',
             '    <tr><td></td><td>E</td><td>F</td></tr>',
             '  </tbody>',
@@ -596,7 +609,9 @@ class ListTableExtensionTest extends TestCase
 
         $expected = implode("\n", [
             '<table>',
-            '  <thead><tr><th scope="col" colspan="2">A</th><th scope="col">C</th></tr></thead>',
+            '  <thead>',
+            '    <tr><th scope="col" colspan="2">A</th><th scope="col">C</th></tr>',
+            '  </thead>',
             '  <tbody>',
             '    <tr><td>x</td><td></td><td>y</td></tr>',
             '  </tbody>',
@@ -625,7 +640,9 @@ class ListTableExtensionTest extends TestCase
 
         $expected = implode("\n", [
             '<table>',
-            '  <thead><tr><th scope="col">H1</th><th scope="col">H2</th></tr></thead>',
+            '  <thead>',
+            '    <tr><th scope="col">H1</th><th scope="col">H2</th></tr>',
+            '  </thead>',
             '  <tbody>',
             '    <tr><td rowspan="2">A</td><td>B</td></tr>',
             '    <tr><td>C</td></tr>',
@@ -828,7 +845,9 @@ class ListTableExtensionTest extends TestCase
 
         $expected = implode("\n", [
             '<table>',
-            '  <thead><tr><th scope="col">H1</th><th scope="col">H2</th></tr></thead>',
+            '  <thead>',
+            '    <tr><th scope="col">H1</th><th scope="col">H2</th></tr>',
+            '  </thead>',
             '  <tbody>',
             '    <tr><td></td><td>x</td></tr>',
             '  </tbody>',
@@ -1020,7 +1039,7 @@ class ListTableExtensionTest extends TestCase
         ]));
 
         $this->assertStringContainsString('<colgroup>', $html);
-        $this->assertStringContainsString('<tfoot><tr>', $html);
+        $this->assertStringContainsString("<tfoot>\n    <tr>", $html);
         $this->assertStringContainsString('text-align: left; vertical-align: top;', $html);
         $this->assertStringNotContainsString('aligns=', $html);
     }